feat: Add support for Firefox and Firefox dev browser
Caveat:
The user will have to install the lz4 compression library as Firefox
stores the bookmarks file in a compressed state using a modified lz4
compression algorithm. The modification is just that the header is a bit
different.

For firefox users, the installation will have to include a run step for
`make` to build the program which will decompress the firefox compressed
file.

Firefox uses profiles and thus, a new option is provided for the profile
name. Default will be `*.default-release` for released version and
`*.dev-edition-default` for developer edition. It is highly recommended
to include the profile name.