feat(firefox): bookmarks are in sqlite database
This is my mistake that I didn't research well. Previously, this would
collect the bookmarks from one of the backup files which is compressed.
Actually, the bookmarks are stored in a sqlite database named
`places.sqlite`.

BREAKING CHANGE: For anyone using this extension with the firefox browser,
it is now dependent on sqlite.lua (https://github.com/tami5/sqlite.lua).
It is NOT required to download the sqlite plugin if you're not using
this with the firefox browser.